requirement.txt
Django
django-cors-headers
dotenv
psycopg2
djangorestframework
djangorestframework_simplejwt
uuid7